--- deep-capitalizer/src/History.js
@@ -1,3 +1,0 @@
-import { createBrowserHistory as history} from 'history';
-
-export default history();
