+++ deep-capitalizer/src/Routes.js
@@ -1,22 +1,27 @@
-
 import React, { Component } from 'react';
 import { Router, Switch, Route } from 'react-router-dom';
 
-import Home from './Home';
-import TestPage from './TestPage';
-import App from './App'
-import history from './History';
+import Home from './components/Home';
+import TestPage from './components/TestPage';
+import App from './components/App';
+import ButtonAppBar from './components/Navbar';
+import { createBrowserHistory } from 'history';
 
-export default class Routes extends Component {
-    render() {
-        return (
-            <Router history={history}>
-                <Switch>
-                    <Route path='/' exact component={Home} />
-                    <Route path='/testPage' exact component={TestPage} />
-                    <Route path='/app' exact component={App} />
-                </Switch>
-            </Router>
-        )
-    }
+export const history = createBrowserHistory();
+
+const Routes = () => {
+  return (
+    <>
+      <ButtonAppBar/>
+      <Router history={history}>
+        <Switch>
+          <Route path='/' exact component={Home} />
+          <Route path='/testPage' exact component={TestPage} />
+          <Route path='/app' exact component={App} />
+        </Switch>
+      </Router>
+    </>
+  )
 }
+
+export default Routes
