+++ deep-capitalizer/src/components/Navbar.js
@@ -5,7 +5,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
-// import MenuIcon from '@material-ui/icons/Menu';
+import MenuIcon from '@material-ui/icons/Menu';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -27,10 +27,10 @@ export default function ButtonAppBar() {
       <AppBar position="static">
         <Toolbar>
           <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-            
+            <MenuIcon/>
           </IconButton>
           <Typography variant="h6" className={classes.title}>
-            News
+            Deep Capitalizer
           </Typography>
           <Button color="inherit">Login</Button>
         </Toolbar>
